Texture image dataset suitable for 3D printing
This image dataset compiles a texture grayscale image containing over 300 repeated units, with pixels set to 192 * 192 and grayscale values ranging from 0 to 255. It is convenient for mapping grayscale images to heights and can be used for 3D printing. The images are collected from the internet, such as texture normal maps and rhinoceros models.